+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added `emr-serverless` module with unit-tests
 - added workflow entries to all IDF modules
 - made `requirements.txt` file of MWAA configurable via a user defined entry from module manifest file
+- added `app-registry` module for being able to scrape app-specific CloudFormation stacks for AWS Solutions
 
 ### **Changed**